Manually set proc.yield in delay callbacks
Fixes the failing delay/randdelay tests where the (success) callback passed to FiberRuntime#run gets called before the delay setTimeout() callbacks got even the chance to execute. This problems stems from the fact that proc.yielded is only set in InputPort and OutputPort when waiting for input or sending, but not when calling Fiber.yield() only.