Fixes #12
The actual problem was that yielded processes (such as delay and
randdelay) were run although they had still been in their sleep time
(i.e. cbpending=true). How was this possible? The OutportPort of an
upstream process connected to delay pushed (and still does) the
downstream process (=delay) onto the queue. The FiberRuntime now checks
on every tick whether a process is currently running callback. In this
case it won't resume the process. Eventual resuming is guaranteed by
FiberRuntime#runAsyncCallback which pushes the process onto the queue
after the callback has completed.